BackendAPI/TranscribeAPI/requirements.txt
google-cloud-storage
ffmpeg-python
google-cloud-speech
pydub

